Implement form for more detailed info for grain
Advanced users may want to provide more information
about their grain. Capturing this data systematically
is better than the old freeform textbox and this change
will make it easier to compare grains.

The extra fields are hidden so that a beginner user is
not overwhelmed. Subcategories are filter based on the
grain selected.

Hooking up the model will be handled in a future commit.