Hotfix memberinfoid vs userid db sync
A user can be created without a memberinfo being created if the
form errors. This has caused the memberinfo id and the user id
to be out of sync on the live database.

TODO plan on making a consistent choice to use either memberinfo id or
userid through out codebase.